+++ Tests/Doctrine/Fixtures/FixtureTest.php
@@ -21,7 +21,8 @@ class FixtureTest extends IntegrationTestCase {
 	/**
 	 * @dataProvider getPlatformConfigs
 	 */
-	public function testImportByUsingFixtureDirectly($platform, $config, $requiredExtension) {
+	public function testImportByUsingFixtureDirectly($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => $platform, 'config' => $config]);
 
 		// [A] add some data which is meant to be removed by importing new data
@@ -42,39 +43,10 @@ class FixtureTest extends IntegrationTestCase {
 	}
 
 	/**
-	 * @dataProvider getPlatformConfigs
-	 */
-	public function testImportByDoctrineFixturesBundle2Command($platform, $config, $requiredExtension) {
-		if (class_exists(SymfonyFixturesLoader::class)) {
-			$this->markTestSkipped('DoctrineFixturesBundle >= 3.0 does not allow loading fixtures from a directory anymore.');
-		}
-
-		$this->initClient($requiredExtension, ['environment' => $platform, 'config' => $config]);
-
-		// [A] add some data which is meant to be removed by importing new data
-		$this->persistGeoPostalCode('DE', '14473', 52.392759, 13.065135);
-
-		// [B] import new data by using a command to load the fixture
-		$application = new Application(static::$kernel);
-		$application->setAutoExit(false);
-		$output = self::executeCommand($application, 'doctrine:fixtures:load', ['--append' => null, '--fixtures' => 'Tests/Doctrine/Fixtures/CraueGeo']);
-		$this->assertEquals(sprintf("  > loading %s\n 177\n", PuertoRicoGeonamesPostalCodeData::class), $output);
-
-		// [A] verify that old data has been removed
-		$this->assertCount(0, $this->getRepo()->findBy(['country' => 'DE']));
-
-		// [B] verify that new data was imported as expected
-		$this->assertCount(177, $this->getRepo()->findAll());
-	}
-
-	/**
 	 * @dataProvider dataImportByDoctrineFixturesBundle3Command
 	 */
-	public function testImportByDoctrineFixturesBundle3Command($platform, $config, $requiredExtension) {
-		if (!interface_exists(FixtureGroupInterface::class)) {
-			$this->markTestSkipped('DoctrineFixturesBundle < 3.1 does not support fixture groups.');
-		}
-
+	public function testImportByDoctrineFixturesBundle3Command($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => 'fixtureAsAService_' . $platform, 'config' => $config]);
 
 		// [A] add some data which is meant to be removed by importing new data
@@ -93,7 +65,8 @@ class FixtureTest extends IntegrationTestCase {
 		$this->assertCount(177, $this->getRepo()->findAll());
 	}
 
-	public function dataImportByDoctrineFixturesBundle3Command() {
+	public static function dataImportByDoctrineFixturesBundle3Command(): array
+    {
 		return self::duplicateTestDataForEachPlatform([
 			[],
 		], 'config_fixtureAsAService.yml');
@@ -102,11 +75,8 @@ class FixtureTest extends IntegrationTestCase {
 	/**
 	 * @dataProvider dataImportByDoctrineFixturesBundle3CommandWithAutoRegistration
 	 */
-	public function testImportByDoctrineFixturesBundle3CommandWithAutoRegistration($platform, $config, $requiredExtension) {
-		if (!interface_exists(FixtureGroupInterface::class)) {
-			$this->markTestSkipped('DoctrineFixturesBundle < 3.1 does not support fixture groups.');
-		}
-
+	public function testImportByDoctrineFixturesBundle3CommandWithAutoRegistration($platform, $config, $requiredExtension): void
+    {
 		$this->initClient($requiredExtension, ['environment' => 'fixtureAsAService_autoRegistration_' . $platform, 'config' => $config]);
 
 		// [A] add some data which is meant to be removed by importing new data
@@ -125,13 +95,15 @@ class FixtureTest extends IntegrationTestCase {
 		$this->assertCount(177, $this->getRepo()->findAll());
 	}
 
-	public function dataImportByDoctrineFixturesBundle3CommandWithAutoRegistration() {
+	public function dataImportByDoctrineFixturesBundle3CommandWithAutoRegistration(): array
+    {
 		return self::duplicateTestDataForEachPlatform([
 			[],
 		], 'config_fixtureAsAService_autoRegistration.yml');
 	}
 
-	private static function executeCommand(Application $application, $command, array $options = []) {
+	private static function executeCommand(Application $application, string $command, array $options = []): string
+    {
 		$options = array_merge($options, [
 			'--env' => $application->getKernel()->getEnvironment(),
 			'--no-debug' => null,